Moved methods to ClassDumper
